Refactor DTO organization and optimize queries with caching (#148)
