+++ src/main/java/com/igniscore/api/dto/client/ClientQueryDTO.java
@@ -1,4 +1,4 @@
-package com.igniscore.api.dto;
+package com.igniscore.api.dto.client;
 
 import com.igniscore.api.model.Client;
