+++ src/main/java/com/igniscore/api/dto/dashboard/MonthlySalesDTO.java
@@ -1,4 +1,4 @@
-package com.igniscore.api.dto;
+package com.igniscore.api.dto.dashboard;
 
 import java.math.BigDecimal;
