+++ src/main/java/com/igniscore/api/dto/product/ProductQueryDTO.java
@@ -1,4 +1,4 @@
-package com.igniscore.api.dto;
+package com.igniscore.api.dto.product;
 
 import com.igniscore.api.model.Product;
