+++ src/main/java/com/igniscore/api/dto/sale/CreateSaleItemDTO.java
@@ -1,4 +1,4 @@
-package com.igniscore.api.dto;
+package com.igniscore.api.dto.sale;
 
 import java.math.BigDecimal;
