+++ src/main/java/com/igniscore/api/repository/ExpirationRepository.java
@@ -1,6 +1,6 @@
 package com.igniscore.api.repository;
 
-import com.igniscore.api.dto.ExpirationProjectionDTO;
+import com.igniscore.api.dto.expiration.ExpirationProjectionDTO;
 import com.igniscore.api.model.Company;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
@@ -12,7 +12,7 @@ import java.util.List;
 public interface ExpirationRepository extends JpaRepository<Company, Integer> {
 
     @Query("""
-        SELECT new com.igniscore.api.dto.ExpirationProjectionDTO(
+        SELECT new com.igniscore.api.dto.expiration.ExpirationProjectionDTO(
             s.id,
             c.name,
             s.date,
@@ -29,7 +29,7 @@ public interface ExpirationRepository extends JpaRepository<Company, Integer> {
     );
 
     @Query("""
-    SELECT new com.igniscore.api.dto.ExpirationProjectionDTO(
+    SELECT new com.igniscore.api.dto.expiration.ExpirationProjectionDTO(
         s.id,
         c.name,
         s.date,
@@ -49,7 +49,7 @@ public interface ExpirationRepository extends JpaRepository<Company, Integer> {
     );
 
     @Query("""
-    SELECT new com.igniscore.api.dto.ExpirationProjectionDTO(
+    SELECT new com.igniscore.api.dto.expiration.ExpirationProjectionDTO(
         s.id,
         c.name,
         s.date,
@@ -69,7 +69,7 @@ public interface ExpirationRepository extends JpaRepository<Company, Integer> {
     );
 
     @Query("""
-    SELECT new com.igniscore.api.dto.ExpirationProjectionDTO(
+    SELECT new com.igniscore.api.dto.expiration.ExpirationProjectionDTO(
         s.id,
         c.name,
         s.date,
